% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_additionalLinters.R
\name{exceptInComments}
\alias{exceptInComments}
\title{First part of a regular expression to exclude the pattern if it appears in
a comment}
\usage{
exceptInComments()
}
\description{
Must be pasted together with the actual pattern that should be
found only outside comment (and roxygen comments).
}
