% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_computeMode.R
\name{mode}
\alias{mode}
\title{Mode}
\usage{
mode(x, notUniqueVal = "notUnique", naVal = "allNA",
  emptyVal = "noValues")
}
\arguments{
\item{x}{Vector}

\item{notUniqueVal}{character: return value if mode is not unique}

\item{naVal}{character: return value if all values in \code{x} are \code{NA}}

\item{emptyVal}{character: return value if \code{x} is empty}
}
\value{
Character vector
}
\description{
Compute mode (most frequent value). The following cases are
covered: empty input, only \code{NA} values, all values are equal, mode not
unique (i.e., several values appear with the same frequence).
}
\examples{
mode(c(1, 1, 2, 3))
mode(1:5)
mode(1:5, 5)
mode(NA, NA, NA)
}
